refactor: change package of actual variables to match the expect
